add security in journalentry

+++ src/main/java/com/example/journal/service/JournalEntryService.java
@@ -36,7 +36,7 @@ public class JournalEntryService {
             journalEntity.setUser(user);
             journalEntity.setDate(LocalDateTime.now()); // fixed: not static
             user.getJournalEntries().add(journalEntity);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }
     }
 
@@ -48,12 +48,27 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(int id, String userName) {
-        User user = userService.findByUserName(userName);
-        if (user != null) {
-            user.getJournalEntries().removeIf(entry -> entry.getId() == id);
-            userService.saveEntry(user);
-            journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(int id, String userName) {
+        boolean remove = false;
+        try {
+            User user = userService.findByUserName(userName);
+            if (user != null) {
+                 remove = user.getJournalEntries().removeIf(entry -> entry.getId() == id);
+                if(remove){
+                    userService.saveUser(user);
+                    journalEntryRepository.deleteById(id);
+                }
+            }
         }
+        catch (Exception e)
+        {
+            System.out.println(e);
+            throw new RuntimeException("An Error occurs while deleting the journal entry",e);
+        }
+
+        return remove;
+
     }
+
 }
